feat(sites-components): port over LexicalRichText component from react-components (#18)
This PR copies over the `LexicalRichText` components from `react-components` without any functionality changes. The actual logic for the component wasn't changed, but some refactoring was done to match the structure of this repo:
- split out types and methods into separate files to follow how other components are laid out
- co-located the Storybook stories and fixtures with the component file
- updated the types from Storybook since this repo uses v7 instead of v6
- added file extensions to imports
- added a default React import to each TSX file

This repo does not use Tailwind, so the component doesn't have any styling when viewed in the Storybook site. A future PR will change all the styling from Tailwind to CSS/SCSS.

J=SLAP-2992
TEST=manual

- Tested that a story for the component is added to the Storybook site, even though it isn't styled
- `npm pack` the `sites-components` package and use it in a test-site with Tailwind. Saw that the component is rendered as expected with styling matching that in the `react-components` Storybook site