feat(sites-components): (no) default styles for address (#19)
For the address component, the only default style we added removes the underline for the `<abbr>` tag. We do not want to enforce any font or text color/size since the user will almost always want to apply their own branding.  The spacing was also considered, but did not need changing.

![image](https://github.com/yext/js/assets/59857107/ebdb06c8-0f6a-4c5c-b614-99016650a237)

J=SLAP-2994
TEST=manual