feat(sites-components): add LexicalRichText component (#26)
Add the `LexicalRichText` component from `@yext/react-components` with some refactoring to match how this repo is structured. Update its styling to match alpha's `RichTextV2Display` component, which uses CSS rather than Tailwind.

A future PR will re-sync the functionality of the component with the one in alpha.

J=SLAP-2992, SLAP-2995
TEST=manual

Add a Storybook story for the component and see that it looks similar to how rich text is rendered in `@yext/react-components` and in the platform. Also, use the component successfully in a test site with styling appearing correctly when the `style.css` bundle is imported.