feat(sites-components): re-sync LexicalRichText with alpha (#24)
Update `LexicalRichText` to more closely match T-Bone's `RichTextV2Display` component in alpha. Specifically, this PR adds support for the `HorizontalRuleNode`, some plugins provided by Lexical, as well as some custom plugins:
- plugin for code highlighting that adds line numbers to code blocks
- plugin for images, although it isn't used for display purposes
- plugin for indenting lists, although it's only for edit mode

Since T-Bone has the image and list indenting plugins in `RichTextV2Display`, we decided to add them here to make it easier in case they decide to update those plugins to affect rendering in the future. `RichTextV2Display` also includes a custom node type for embedded keys, but since this node type would only ever be present in platform, it doesn't need to be added to this component. See thread [here](https://yext.slack.com/archives/CNVUY09HS/p1699984807893229?thread_ts=1699471302.805499&cid=CNVUY09HS) for both of these discussions.

Even with these changes, there are small differences in styling between Storm and this component. This is due to the fact that all Storm pages have some shared default CSS applied to them, such as a font size, which affects the rendering of `RichTextV2Display`. For now, we're not adding the Storm styling to this component, but that may change in the future.

J=SLAP-3006, SLAP-3007
TEST=manual

In a test site that renders a rich text v2 field, see that horizontal rule nodes are handled successfully and that styling matches up with Entity Edit's in platform with a couple exceptions like font size for lists (see above).