chore(sites-components): automated-linting (#23)
This PR adds automated linting to this repo.  The eslint config was changed to extend `@yext/eslint-config/typescript-react` to match Yext linting standards.  If autofixes are made and pushed, then the test will fail and the test will run again on the new commit.  Concurrency checking was also implemented to make sure that when tests for a new commit are run, the tests for the old commit will halt.

Since sites-components does not currently comply with the current eslint standards, linting many errors surfaced.  We would like to provide eslint and prettier autofixes, but also not block PRs.  So, this new workflow will not have a required PR check. Note that the `automated-linting` check is expected to fail, since we have not addressed the linting errors.

J=SLAP-3003
TEST=auto,manual