chore(sites-components): add visual regression testing (#31)
This PR adds visual regression testing to sites-components. To save and compare screenshots, [jest-image-snapshot](https://github.com/americanexpress/jest-image-snapshot) was used.  A workflow was added to automatically update screenshots for PRs.  Like other automated workflows, if this PR pushes its changes, it will cancels past workflows and trigger a new workflow. We should be cognizant that if this workflow pushes snapshot updates, we have made a UI change that should be verified.

Note that this PR adds snapshots for all components.  This includes components (e.g. the map component) that are currently broken in Storybook, and components which we have not yet decided on their default styles.

After this PR is merged, we will create a required check for this test.

J=SLAP-3013
TEST=auto,manual