revert(sites-components): remove Hours export (#32)
This PR manually reverts the export of the `Hours` component added as part of PR #21. Currently, CSS for the `Hours` component causes errors if the CSS bundle is imported in a repo with Tailwind. Since the `Hours` component wasn't exported previously, removing it won't be a breaking change. And in the meantime, the other exported components can still be used along with their styling.

J=none
TEST=manual

See that a test site with Tailwind can now import `@yext/sites-components/style.css` without errors and have the styling applied.